backend/re_bubble_engine/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.1
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[asyncio]==5.2.0
httpx==0.27.2
numpy==2.1.3
pandas==2.2.3
google-generativeai==0.8.3
groq==0.12.0
openai==1.54.0
langgraph==0.2.45
langchain-core==0.3.18
langchain-community==0.3.8
qdrant-client==1.12.1
sentence-transformers==3.3.1
apscheduler==3.10.4
structlog==24.4.0
tenacity==9.0.0
sse-starlette==2.1.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
